[FIX] refactor config package into separate files
